add some react logic

+++ src/main/java/ru/yandex/practicum/intershop/dto/ItemResponseDto.java
@@ -8,10 +8,10 @@ import java.math.BigDecimal;
 @Data
 @Accessors(chain = true)
 public class ItemResponseDto {
-    private final Integer id;
+    private final Long id;
     private final String title;
     private final BigDecimal price;
     private final String description;
     private final String imgPath;
-    private int count = 0;
+    private Integer count = 0;
 }
